% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by_region.R
\name{aggregate_regions}
\alias{aggregate_regions}
\title{Aggregate positional methylation data by regions provided with an annotation file.}
\usage{
aggregate_regions(modseq_dat, annot_file)
}
\arguments{
\item{modseq_dat}{A duckdb object of methylation data already processed with summarize_by_pos().}

\item{annot_file}{Path to an annotation file of regions to aggregate data into- TSV format.
This file is required to have three columns- chrom, start, end. Optional fourth column of region_name can be included.}
}
\value{
A duckdb object of all regional methylation data to be processed.
}
\description{
Aggregate positional methylation data by regions provided with an annotation file.
}
\examples{
regional_data = aggregate_regions(data, "CpG_islands.tsv")
}
