% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ch3_diff.R
\name{calc_ch3_diff}
\alias{calc_ch3_diff}
\title{Calculate Differential Methylation}
\usage{
calc_ch3_diff(
  ch3_db,
  call_type = "positions",
  cases,
  controls,
  mod_type = "mh",
  calc_type = NULL
)
}
\arguments{
\item{ch3_db}{A list containing the database file path. This should be a valid "ch3_db" class object.}

\item{call_type}{A string representing the name of the table in the database from which to pull the data. 
Default is "positions".}

\item{cases}{A character vector containing the sample names for the case group.}

\item{controls}{A character vector containing the sample names for the control group.}

\item{mod_type}{A string indicating the type of modification to analyze. 
Default is "mh" for methylation/hydroxymethylation.}

\item{calc_type}{A string specifying the statistical method to use for calculating p-values. 
Options include "wilcox", "fast_fisher", "r_fisher", and "log_reg". 
Default is NULL, in which case "wilcox" is used if there are replicates in either
group, otherwise "fast_fisher" is used.}
}
\value{
A list containing the updated "ch3_db" object with the latest tables in the database, including "meth_diff".
}
\description{
This function calculates differential methylation between specified case and control groups using various statistical methods. 
The results are stored in a DuckDB database for further analysis.
}
\details{
The function connects to the specified DuckDB database and retrieves methylation data from the specified call type table. 
It summarizes the data for cases and controls, calculates p-values based on the specified method, and stores the results in the 
"meth_diff" table.
}
\examples{
 # Specify the path to the database
 ch3_db <- system.file("my_data.ch3.db", package = "MethylSeqR")
 
 # Get methylation statistics for the 'positions' call type without plotting
 calc_ch3_diff(ch3_db = ch3_db, 
               call_type = "positions",
               cases = c("Blood1_chr21", "Blood2_chr21", "Blood3_chr21"),
               controls = c("Sperm1_chr21", "Sperm2_chr21", "Sperm3_chr21")))
               
}
