% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_diff.R
\name{calc_mod_diff}
\alias{calc_mod_diff}
\title{Calculate Differential Methylation}
\usage{
calc_mod_diff(
  modseq_dat,
  cases,
  controls,
  mod_type = "mh",
  calc_type = "fast_fisher"
)
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data.}

\item{cases}{A vector of sample names to be treated as cases.}

\item{controls}{A vector of sample names to be treated as controls.}

\item{mod_type}{A string specifying the type of methylation data, either "mh" or "m".}

\item{calc_type}{A string specifying the calculation type, either
"fast_fisher", "r_fisher", or "log_reg".}
}
\value{
A data frame with p-values and differential methylation statistics.
}
\description{
This function calculates differential methylation between cases and controls
using either Fisher's exact test or logistic regression.
}
\examples{
\dontrun{
calc_mod_diff(modseq_dat, cases, controls)
calc_mod_diff(modseq_dat, cases, controls, mod_type = "m", calc_type = "log_reg")
}

}
