% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_modseq.R
\name{cluster_modseq}
\alias{cluster_modseq}
\title{Cluster Methylation Data}
\usage{
cluster_modseq(ch3_db, call_type = c("positions"))
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data. The data frame can
either contain positional or regional data. If it contains regional data, it must
have a column named \code{region_name}.}
}
\value{
None. The function plots a dendrogram.
}
\description{
This function clusters methylation data using hierarchical clustering
and plots a dendrogram to visualize the clustering.
}
\examples{
\dontrun{
cluster_modseq(modseq_dat)
}

}
