% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_modseq.R
\name{cor_modseq}
\alias{cor_modseq}
\title{Compute and Optionally Plot Correlation Matrix of Methylation Data}
\usage{
cor_modseq(modseq_dat, plot = FALSE)
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data. The data frame can
either contain positional or regional data. If it contains regional data, it must
have a column named \code{region_name}.}

\item{plot}{A logical value indicating whether to plot the correlation matrix.
Default is \code{FALSE}.}
}
\value{
The correlation matrix of the methylation data.
}
\description{
This function computes the Pearson correlation matrix of methylation data
and optionally plots the correlation matrix.
}
\examples{
\dontrun{
correlation_matrix <- cor_modseq(modseq_dat, plot = TRUE)
}

}
