% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_modseq.R
\name{density_modseq}
\alias{density_modseq}
\title{Generate and Plot Density of Methylation Data}
\usage{
density_modseq(modseq_dat, samples = c("all"), save = FALSE)
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data. The data frame can
either contain positional or regional data. If it contains regional data, it must
have a column named \code{region_name}.}

\item{samples}{A character vector specifying which samples to compare. If set to \code{"all"},
all samples in the data will be compared. If fewer than two samples are specified, the function will
stop execution with an error message.}

\item{save}{A logical value indicating whether to save the plots to PNG files. If \code{TRUE},
the plots will be saved with filenames based on the sample names.}
}
\value{
A list of ggplot objects representing the density plots for each pair of samples, if \code{save}
is \code{FALSE}. If \code{save} is \code{TRUE}, the plots are saved as PNG files and the function
returns \code{NULL}.
}
\description{
This function generates density plots for pairs of methylation samples
and visualizes them using raster plots.
}
\examples{
\dontrun{
density_modseq(modseq_dat)
density_modseq(your_data, samples = c("barcode81", "barcode84", "barcode87"))
}

}
