% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sliding_window.R
\name{.compute_windows_chr}
\alias{.compute_windows_chr}
\title{Compute Sliding Windows for All Chromosomes}
\usage{
.compute_windows_chr(cur_chrom, db_con, modseq_dat, window_size)
}
\arguments{
\item{modseq_dat}{A DuckDB object of methylation data already processed.}

\item{window_size}{The size of the sliding window. Default is 500.}
}
\value{
A DuckDB table of sliding windows.
}
\description{
This function sets up a database, iterates over all chromosomes,
calls `.compute_windows_chr` for each, and returns the final table
of sliding windows.
}
\examples{
sliding_windows <- compute_sliding_windows(modseq_dat, 500)

}
