% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_closeDB.R
\name{.helper_closeDB}
\alias{.helper_closeDB}
\title{Close Database Connection}
\usage{
.helper_closeDB(ch3_db)
}
\arguments{
\item{ch3_db}{An object containing information about the database, including the list of tables.}

\item{db_con}{A database connection object. This should be a connection created by a database connection function.}
}
\value{
None. This function is called for its side effects (updating the `ch3_db` object and closing the connection).
}
\description{
This function updates the table list in the `ch3_db` object and closes the database connection.
}
\details{
The function updates the `tables` attribute of the `ch3_db` object with the current list of tables in the connected database 
before closing the connection.
}
