% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_connectDB.R
\name{.helper_connectDB}
\alias{.helper_connectDB}
\title{Connect to a Database}
\usage{
.helper_connectDB(ch3_db)
}
\arguments{
\item{ch3_db}{A character string representing the file path to the DuckDB database or an object of class `ch3_db`.}
}
\value{
A database connection object.
}
\description{
This internal function establishes a connection to a DuckDB database. It can handle both a character file name 
or an object of class `ch3_db` to open the database.
}
\details{
This function checks the class of `ch3_db` and attempts to connect to the database. If `ch3_db` is a character string, 
it will create an object of class `ch3_db`. If `ch3_db` is already of class `ch3_db`, it will directly establish a 
connection to the database.
}
\note{
This function is intended for internal use within the package.
}
\keyword{internal}
