% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_purgeTables.R
\name{.helper_purgeTables}
\alias{.helper_purgeTables}
\title{Internal Function to Purge Unwanted Tables from Database}
\usage{
.helper_purgeTables(db_con, keep_tables = c("positions"))
}
\arguments{
\item{keep_tables}{A character vector of table names to keep in the database. Defaults to keeping only 
the "positions" table.}

\item{ch3_db}{A character string or an object of class `ch3_db` representing the DuckDB database to connect to.}
}
\value{
None. This function is called for its side effects (modifying the database).
}
\description{
This internal function connects to a DuckDB database and removes tables that are not specified in the 
`keep_tables` list. It retains only the tables that you want to keep in the database.
}
\details{
The function connects to the specified database, lists all tables, and removes those not included in 
the `keep_tables` vector. After purging, it prints the names of the remaining tables in the database.
}
\keyword{internal}
