% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_names.R
\name{get_column_names}
\alias{get_column_names}
\title{Get Column Names from a DuckDB Table}
\usage{
get_column_names(ch3_db, table_name)
}
\arguments{
\item{ch3_db}{Path to the DuckDB database file (e.g., `"my_data.ch3.db"`).}

\item{table_name}{Name of the table whose column names are to be retrieved.}
}
\value{
A character vector of column names from the specified table.
}
\description{
Returns a character vector of column names from a specified table
in a DuckDB database file.
}
\examples{
\dontrun{
get_column_names("my_data.ch3.db", "windows")
}

}
