% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cov_stats.R
\name{get_cov_stats}
\alias{get_cov_stats}
\title{Calculate and Plot Coverage Statistics}
\usage{
get_cov_stats(ch3_db, call_type = c("positions", "regions"), plot = FALSE)
}
\arguments{
\item{ch3_db}{A data base either linking to the file name or of class ch3_db.}

\item{call_type}{Either positions or regions data to analyze coverage on.}

\item{plot}{Logical, if \code{TRUE}, the function will generate a histogram of
the coverage data. Default is \code{FALSE}.}
}
\value{
If \code{plot} is \code{FALSE}, the function prints summary statistics
and percentiles of the coverage data. If \code{plot} is \code{TRUE}, it prints a
histogram of the log-transformed coverage data.
}
\description{
This function calculates and optionally plots statistics for coverage data from
methylation sequencing experiments. It can handle both positional and regional
methylation data.
}
\examples{
\dontrun{
get_cov_stats(modseq_dat)
get_cov_stats(modseq_dat, plot = TRUE)
}

}
