% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_stats.R
\name{get_db_stats}
\alias{get_db_stats}
\title{Get Database Statistics}
\usage{
get_db_stats(ch3_db)
}
\arguments{
\item{ch3_db}{Path to the `.ch3.db` file or a `ch3_db` object.}
}
\value{
Invisibly returns a list of stats from the database.
}
\description{
Prints out a summary of the ch3 database, including size, tables, and unique sample names.
}
