% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mod_stats.R
\name{get_mod_stats}
\alias{get_mod_stats}
\title{Calculate and Plot Methylation Statistics}
\usage{
get_mod_stats(modseq_dat, plot = FALSE)
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data. It should have one of
the following columns: \code{mh_frac}, \code{mean_mh_frac}, \code{m_frac}, or \code{mean_m_frac}.}

\item{plot}{Logical, if \code{TRUE}, the function will generate a histogram of
the methylation data. Default is \code{FALSE}.}
}
\value{
If \code{plot} is \code{FALSE}, the function prints summary statistics
and percentiles of the methylation data. If \code{plot} is \code{TRUE}, it prints a
histogram of the methylation data.
}
\description{
This function calculates and optionally plots statistics for methylation data from
sequencing experiments. It can handle both positional and regional methylation data.
}
\examples{
\dontrun{
get_mod_stats(modseq_dat)
get_mod_stats(modseq_dat, plot = TRUE)
}

}
