% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Collect Table from DuckDB Database as Tibble}
\usage{
get_table(ch3_db, table_name)
}
\arguments{
\item{ch3_db}{A list containing the database file path. This should be a valid "ch3_db" class object.}

\item{table_name}{A string representing the name of the table to collect from the database.}
}
\value{
A tibble containing the collected data from the specified database table. If the table retrieval fails, an empty tibble is returned.
}
\description{
This function connects to a DuckDB database and collects a specified table as a tibble.
}
\details{
The function establishes a connection to the DuckDB database using \code{.helper_connectDB}.
It retrieves the specified table as a tibble. If an error occurs during table retrieval, 
a message with the error is displayed. The database connection is closed after retrieving 
the data, regardless of success or failure.
}
\examples{
# Assuming ch3_db is a valid database object and "positions" is a table in the database
ch3_db <- system.file("my_data.ch3.db", package = "MethylSeqR")
positions = get_table(ch3_db, "positions")

}
