% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ch3_archive.R
\name{make_ch3_archive}
\alias{make_ch3_archive}
\title{Create CH3 Archive from Methylation Data}
\usage{
make_ch3_archive(file_name, sample_name, out_path)
}
\arguments{
\item{file_name}{A string representing the path to the input file containing methylation data.}

\item{sample_name}{A string representing the name of the sample. This name will be included in 
the output files.}

\item{out_path}{A string representing the path where the output CH3 dataset will be stored.}
}
\value{
A string representing the path of the created archive file, useful for piping or assignment.
The function is designed to be used as a part of a data processing pipeline.
}
\description{
This function reads methylation data from a specified file and creates an archive of the data 
in CH3 format, storing it in a designated output path. It processes the data by selecting relevant 
columns and mutating the reference position based on the modification strand.
}
\details{
The function uses the Arrow package to read the input TSV dataset and select specific columns. 
It then adjusts the reference position based on the modification strand before writing the dataset 
to the specified output path. The output files are named using the provided `sample_name` and a 
unique index.
}
\examples{
make_ch3_archive(file_name = "path/to/data.tsv", 
                 sample_name = "Sample1", 
                 out_path = "path/to/output")

}
