% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pos_db.R
\name{make_pos_db}
\alias{make_pos_db}
\title{Create Position Database from CH3 Files}
\usage{
make_pos_db(
  ch3_files,
  ch3_db,
  min_call_prob = 0.9,
  min_length = 100,
  min_base_qual = 10
)
}
\arguments{
\item{ch3_files}{A string representing the directory path containing CH3 files. 
The function will read all files with the ".ch3" extension in this directory.}

\item{ch3_db}{A string representing the path where the positions database will be created. 
The extension ".ch3.db" will be appended if not already present.}

\item{min_call_prob}{A numeric value representing the minimum call probability threshold. 
Only calls with a probability greater than or equal to this value will be included. Default is 0.9.}

\item{min_length}{A numeric value representing the minimum read length. 
Only reads meeting or exceeding this length will be processed. Default is 100.}

\item{min_base_qual}{A numeric value representing the minimum base quality. 
Only reads with quality scores at or above this threshold will be included. Default is 10.}
}
\value{
A list containing the database file path and the updated list of tables in the database.
This can be useful for further processing or analysis.
}
\description{
This function processes a collection of CH3 files to create a positions database, 
filtering and summarizing the methylation data based on specified criteria. 
The resulting database is stored in a DuckDB format, allowing for efficient queries.
}
\details{
The function connects to a DuckDB database and processes each CH3 file to extract relevant methylation data. 
The data is filtered according to the specified thresholds and summarized by sample name, chromosome, and reference position. 
The final summary statistics, including counts and fractions for methylated and hydroxymethylated bases, are written to the database.
}
\examples{
# Set up the file path for the test data located in inst/test_data/
ch3_files <- system.file("test_data", package = "MethylSeqR")
ch3_db <- tempfile("example_db")

# Run the function with the example data
result <- make_pos_db(ch3_files, ch3_db)
print(result)

}
