% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{MethylSeqR}
\alias{MethylSeqR-package}
\alias{MethylSeqR}
\title{MethylSeqR: Methylation Data Analysis from Nanopore Sequencing}
\description{
MethylSeqR is an R package developed by Wasatch Biolabs for efficient
preprocessing, summarization, and visualization of methylation data.
It supports native DNA methylation sequencing workflows and is
optimized for high-throughput pipelines.
}
\section{Features}{

- Database-backed storage using DuckDB
- Sliding window, positional, regional, and read-level summarization of methylation levels
- Support for multiple methylation contexts
- Quality Control Visualization
- Efficient Differential Methylation Analysis
}

\references{
Wasatch Biolabs (2025). *MethylSeqR: Tools for Methylation Analysis in Clinical and Research Settings.*
https://www.wasatchbiolabs.com/

For bug reports and feature requests:  
https://github.com/Wasatch-Biolabs-Bfx/MethylSeqR
}
\author{
Jonathon Hill \email{jonathon@wasatchbiolabs.com} (aut, cre)
Hailey Zimmerman \email{hailey@renewbt.com} (aut)
}
