% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_db.R
\name{open_dat}
\alias{open_dat}
\title{Open methylation data from a directory}
\usage{
open_dat(
  path,
  samples = "all",
  chrs = "all",
  min_read_length = 100,
  min_call_prob = 0.9,
  min_base_qual = 10,
  max_memory = "2GB",
  max_threads = 4
)
}
\arguments{
\item{path}{Path to database directory}

\item{samples}{Vector of samples to extract methylation data from. Default is all}

\item{chrs}{Vector of chromosomes to extract. Default is all}

\item{min_read_length}{Minimum read length to allow. Filters any read less than this parameter}

\item{min_base_qual}{Minimum base quality}

\item{max_memory}{Maximum memory to be colllected}

\item{max_threads}{Maximum threads to be allowed}
}
\value{
A duckdb object of all methylation data to be processed.
}
\description{
Open methylation data from a directory
}
\examples{
data = open_dat(my_path)
}
