% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_modseq.R
\name{pca_modseq}
\alias{pca_modseq}
\title{Perform Principal Component Analysis (PCA) on Methylation Data}
\usage{
pca_modseq(modseq_dat)
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data. It should include
sample names, methylation fractions, and may include region names.}
}
\value{
A PCA plot showing the first two principal components and the
        variance explained by each component. The PCA summary is also printed
        to the console.
}
\description{
This function performs PCA on methylation data to reduce dimensionality
and visualize the variance explained by the principal components.
}
\examples{
pca_modseq(data)

}
