% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ch3_pca.R
\name{plot_ch3_pca}
\alias{plot_ch3_pca}
\title{Perform PCA on Methylation Data}
\usage{
plot_ch3_pca(
  ch3_db,
  call_type = "positions",
  save_path = NULL,
  max_rows = NULL
)
}
\arguments{
\item{ch3_db}{A list containing the database file path. This should be a valid "ch3_db" class object.}

\item{call_type}{A string representing the name of the table in the database from which to pull the data. 
Default is "positions".}

\item{save_path}{Pathway to save the plot to. Usually .pdf or .png.}

\item{max_rows}{The maximum amount of rows wanted for calculation. This argument can help analysis run faster when there is a lot of data.}
}
\value{
A PCA plot is produced, showing the first two principal components of the methylation data.
The function also prints a summary of the PCA results and the PCA data used for plotting.
}
\description{
This function performs Principal Component Analysis (PCA) on methylation data retrieved from a DuckDB database.
It aggregates the methylation fraction data based on the specified call type and prepares it for PCA analysis.
}
\details{
The function connects to the specified DuckDB database, retrieves the methylation data from the specified call type table, 
and aggregates the data either by regions or chromosomal positions. PCA is then performed on the aggregated data, and 
a scatter plot of the first two principal components is generated.
}
\examples{
 # Specify the path to the database
 ch3_db <- system.file("my_data.ch3.db", package = "MethylSeqR")
 
 # Calculate PCA
 plot_ch3_pca(ch3_db)

}
