% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_modseq.R
\name{qc_modseq}
\alias{qc_modseq}
\title{Quality Control Wrapper for Methylation Data}
\usage{
qc_modseq(modseq_dat, plot = TRUE)
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data. It should include
relevant columns such as sample names, coverage, and methylation fractions.}

\item{plot}{A logical value indicating whether to plot the results. Defaults to TRUE.}
}
\value{
The function performs various quality control analyses and generates plots
        as specified by the `plot` parameter. It does not return any values.
}
\description{
This function performs a series of quality control analyses on methylation data,
including coverage statistics, methylation statistics, correlation analysis,
clustering, PCA, and methylation density estimation.
}
\examples{
qc_modseq(data)

}
