% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pos_analyses.R
\name{run_pos_analyses}
\alias{run_pos_analyses}
\title{Run Positional Analyses on Methylation Data}
\usage{
run_pos_analyses(modseq_dat, cases, controls)
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data. It should include relevant
columns such as sample names, methylation counts, and coverage.}

\item{cases}{A character vector of sample names designated as cases for analysis.}

\item{controls}{A character vector of sample names designated as controls for analysis.}
}
\value{
A data frame containing the results of the positional analyses, including
        differences in methylation between cases and controls, as well as p-values.
}
\description{
This function serves as a wrapper to perform positional analyses on methylation
data by summarizing it by position and calculating the difference in methylation
between specified case and control groups using the fast Fisher's exact test.
}
\examples{
results <- run_pos_analyses(modseq_data, cases = c("case1", "case2"), controls = c("control1", "control2"))

}
