% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regional_analyses.R
\name{run_regional_analyses}
\alias{run_regional_analyses}
\title{Run Regional Analyses Wrapper on Methylation Data}
\usage{
run_regional_analyses(modseq_dat, annot_file, cases, controls)
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data. It should include relevant
columns such as sample names, methylation counts, and coverage.}

\item{annot_file}{A character string representing the path to an annotation file
that specifies the regions to aggregate the methylation data into.
The file should contain columns like 'chrom', 'start', 'end',
and optionally 'region_name'.}

\item{cases}{A character vector of sample names designated as cases for analysis.}

\item{controls}{A character vector of sample names designated as controls for analysis.}
}
\value{
A data frame containing the results of the regional analyses, including
        differences in methylation between cases and controls, as well as p-values.
}
\description{
This function serves as a wrapper to perform regional analyses on methylation
data by summarizing it by position, aggregating the data into specified regions
using an annotation file, and calculating the difference in methylation
between designated case and control groups using the fast Fisher's exact test.
}
\examples{
results <- run_regional_analyses(modseq_data, "regions.tsv", cases = c("case1", "case2"), controls = c("control1", "control2"))

}
