% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ch3.R
\name{sample_ch3}
\alias{sample_ch3}
\title{Downsample a CH3 File to a Desired Coverage}
\usage{
sample_ch3(
  ch3_file,
  out_path,
  coverage = 5,
  sample_name_suffix = "",
  genome_size = 3.2e+09
)
}
\arguments{
\item{ch3_file}{A string representing the file path to the input CH3 file.}

\item{out_path}{A string specifying the directory where the output CH3 file will be saved.}

\item{coverage}{A numeric value specifying the desired coverage for the downsampled CH3 file. Defaults to 5.}

\item{sample_name_suffix}{A string to append to the sample name in the output file. Defaults to an empty string.}

\item{genome_size}{A numeric value representing the size of the genome in base pairs. Defaults to 3.2e9 (human genome size).}
}
\value{
The function does not return a value but writes the downsampled CH3 file to the 
specified output path.
}
\description{
This function downsamples a CH3 file to a specified coverage and writes the 
resulting data to a new CH3 file.
}
\details{
This function calculates the current coverage of the input CH3 file based on the 
genome size and read lengths. If the current coverage is less than the desired coverage, 
the function copies the full file. Otherwise, it samples a fraction of reads to achieve 
the specified coverage. The sampled data is written to a new CH3 file with an updated 
sample name.

The sample name in the output file is generated using the desired coverage, the 
original sample name extracted from the input file path, and an optional suffix.
}
