% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_by_pos.R
\name{summarize_by_pos}
\alias{summarize_by_pos}
\title{Summarize methylation data by reference positions}
\usage{
summarize_by_pos(modseq_dat, score = "mh")
}
\arguments{
\item{modseq_dat}{Methylation data in a duckdb object as opened by open_dat().}

\item{score}{A methylation score type. Default is mh, can also include m, h, or a vector of score types wanted.}
}
\value{
A duckdb object of methylation scores summarized by each position.
}
\description{
Summarize methylation data by reference positions
}
\examples{
summarize_by_pos(data)
summarize_by_pos(data, score = "m")
summarize_by_pos(data, score = c("m", "mh", "h")
}
