% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ch3_regions.R
\name{summarize_ch3_regions}
\alias{summarize_ch3_regions}
\title{Summarize Methylation Data by Regions}
\usage{
summarize_ch3_regions(
  ch3_db,
  table_name = "regions",
  region_file,
  mod_type = c("c", "m", "h", "mh"),
  chrs = c(as.character(1:22), paste0("chr", 1:22), "chrX", "chrY", "chrM", paste0("Chr",
    1:22), "ChrX", "ChrY", "ChrM"),
  min_num_calls = 1
)
}
\arguments{
\item{ch3_db}{A list containing the database file path. This should be a valid "ch3_db" class object.}

\item{table_name}{A string specifying what the user would like the name to be called in the database. Default is "regions".}

\item{region_file}{A string representing the path to the BED or CSV file that contains the region annotations.}

\item{mod_type}{A character vector specifying the modification types to include. Options are  `"c"` (unmodified cytosine),
`"m"` (methylation), `"h"` (hydroxymethylation), 
  and `"mh"` (methylated + hydroxymethylated).}

\item{chrs}{A character vector specifying which chromosomes to include. Default includes all autosomes, 
sex chromosomes (chrX, chrY), and mitochondrial chromosome (chrM).}

\item{min_num_calls}{An integer specifying the minimum number of calls required for inclusion in the summary.
Default is 1.}
}
\value{
The updated `ch3_db` object with the summarized regions data added to the DuckDB database.
}
\description{
This function summarizes methylation data from a DuckDB database based on specified regions 
defined in a BED, TSV, or CSV file. It performs a join operation between the methylation data and 
the regions specified in the annotation file, allowing for different types of joins.
}
\details{
The function reads the region annotations from the regional annotation file and checks for its validity.
It connects to the DuckDB database, creates a summarized table of methylation data based on the specified 
regions, and performs the join operation according to the specified join type. A progress bar is displayed 
during the summarization process. The resulting data is stored in a table called "regions" within the database.
}
\examples{
 # Specify the path to the database
 ch3_db <- system.file("my_data.ch3.db", package = "MethylSeqR")
 region_bed = system.file("Islands_hg38_test.csv", package = "MethylSeqR")
 
 # Summarize Regions using annotation table
 summarize_ch3_regions(ch3_db, region_bed)

}
