% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_regions.R
\name{summarize_regions}
\alias{summarize_regions}
\title{Summarize Methylation Data by Regions}
\usage{
summarize_regions(ch3_db, region_bed, join_type = "inner")
}
\arguments{
\item{ch3_db}{A list containing the database file path. This should be a valid "ch3_db" class object.}

\item{region_bed}{A string representing the path to the BED file that contains the region annotations.}

\item{join_type}{A string indicating the type of join to perform. Options are "inner", "left", 
"right", or "full". Default is "inner".}
}
\value{
The updated `ch3_db` object with the summarized regions data added to the DuckDB database.
}
\description{
This function summarizes methylation data from a DuckDB database based on specified regions 
defined in a BED file. It performs a join operation between the methylation data and the regions 
specified in the annotation file, allowing for different types of joins.
}
\details{
The function reads the region annotations from the specified BED file and checks for its validity.
It connects to the DuckDB database, creates a summarized table of methylation data based on the specified 
regions, and performs the join operation according to the specified join type. A progress bar is displayed 
during the summarization process. The resulting data is stored in a table called "regions" within the database.
}
\examples{
 # Specify the path to the database
 ch3_db <- system.file("my_data.ch3.db", package = "MethylseqR")
 
 region_bed = system.file("Islands_hg38_ucsc.csv", package = "MethylseqR")
 # Summarize Regions using anotation table
 summarize_regions(ch3_db, region_bed)

}
