% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_windows.R
\name{summarize_windows}
\alias{summarize_windows}
\title{Summarize Methylation Data using a Sliding Window}
\usage{
summarize_windows(
  ch3_db,
  window_size = 1000,
  step_size = 10,
  mod_type = c("c", "m", "h", "mh"),
  chrs = c(as.character(1:22), paste0("chr", 1:22), "chrX", "chrY", "chrM", paste0("Chr",
    1:22), "ChrX", "ChrY", "ChrM"),
  min_cov = 1,
  overwrite = TRUE
)
}
\arguments{
\item{ch3_db}{A list containing the database file path. This should be a valid "ch3_db" class object.}

\item{window_size}{An integer specifying the size of the sliding window in base pairs. Default is 1000.}

\item{step_size}{An integer specifying the number of base pairs to step forward with each window. Default is 10.}

\item{mod_type}{A character vector specifying the modification types to include. Options are  `"c"` (unmodified cytosine),
`"m"` (methylation), `"h"` (hydroxymethylation), 
  and `"mh"` (methylated + hydroxymethylated).}

\item{chrs}{A character vector specifying which chromosomes to include. Default includes all autosomes, 
sex chromosomes (chrX, chrY), and mitochondrial chromosome (chrM).}

\item{overwrite}{A logical indicating whether to overwrite the existing "windows" table if it exists. Default is TRUE.}

\item{call_type}{A string indicating the type of data to summarize. Default is "positions".}
}
\value{
The updated `ch3_db` object with the summarized windows data added to the DuckDB database.
}
\description{
This function summarizes methylation data from a DuckDB database by creating 
sliding windows over the specified genomic regions. It allows for the adjustment 
of window size and step size to control the granularity of the summarization.
}
\details{
The function connects to a DuckDB database and removes any existing "windows" and "temp_table" tables if necessary. 
It creates a sequence of offsets based on the specified window and step sizes, and then it iterates through 
these offsets to generate sliding windows of methylation data. A progress bar is displayed during the operation.

The function utilizes the helper function `.make_window` to perform the actual window calculation and summarization.
The resulting summarized data is stored in a table called "windows" within the database.
}
\examples{
 # Specify the path to the database
 ch3_db <- system.file("example_ch3.ch3.db", package = "MethylSeqR")
 
 # Calculate windows
 summarize_windows(ch3_db, window_size = 100, step_size = 100)

}
