% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_diff.R
\name{calc_mod_diff}
\alias{calc_mod_diff}
\title{Calculate Differential Methylation}
\usage{
calc_mod_diff(
  ch3_db,
  call_type = "positions",
  cases,
  controls,
  mod_type = "mh",
  calc_type = "fast_fisher"
)
}
\arguments{
\item{ch3_db}{A list containing the database file path. This should be a valid "ch3_db" class object.}

\item{call_type}{A string representing the name of the table in the database from which to pull the data. 
Default is "positions".}

\item{cases}{A character vector containing the sample names for the case group.}

\item{controls}{A character vector containing the sample names for the control group.}

\item{mod_type}{A string indicating the type of modification to analyze. 
Default is "mh" for methylation/hydroxymethylation.}

\item{calc_type}{A string specifying the statistical method to use for calculating p-values. 
Options include "fast_fisher", "r_fisher", and "log_reg". Default is "fast_fisher".}
}
\value{
A list containing the updated "ch3_db" object with the latest tables in the database, including "meth_diff".
}
\description{
This function calculates differential methylation between specified case and control groups using various statistical methods. 
The results are stored in a DuckDB database for further analysis.
}
\details{
The function connects to the specified DuckDB database and retrieves methylation data from the specified call type table. 
It summarizes the data for cases and controls, calculates p-values based on the specified method, and stores the results in the 
"meth_diff" table.
}
\examples{
 # Specify the path to the database
 ch3_db <- system.file("my_data.ch3.db", package = "MethylSeqR")
 
 # Get methylation statistics for the 'positions' call type without plotting
 calc_mod_diff(ch3_db = ch3_db, 
               call_type = "positions",
               cases = c("chr21_5xSample_BloodA", "chr21_5xSample_BloodE", "chr21_5xSample_BloodT"),
               controls = c("chr21_5xSample_SpermA", "chr21_5xSample_SpermB", "chr21_5xSample_SpermD"))

}
