% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_modseq.R
\name{density_modseq}
\alias{density_modseq}
\title{Generate and Plot Density of Methylation Data}
\usage{
density_modseq(modseq_dat)
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data. The data frame can
either contain positional or regional data. If it contains regional data, it must
have a column named \code{region_name}.}
}
\value{
None. The function prints density raster plots for pairs of methylation samples.
}
\description{
This function generates density plots for pairs of methylation samples
and visualizes them using raster plots.
}
\examples{
\dontrun{
density_modseq(modseq_dat)
}

}
