% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cov_stats.R
\name{get_cov_stats}
\alias{get_cov_stats}
\title{Calculate and Plot Coverage Statistics}
\usage{
get_cov_stats(modseq_dat, plot = FALSE)
}
\arguments{
\item{modseq_dat}{A data frame containing methylation data. It should have a
column named \code{cov} for positional data or \code{mean_cov} for regional data.}

\item{plot}{Logical, if \code{TRUE}, the function will generate a histogram of
the coverage data. Default is \code{FALSE}.}
}
\value{
If \code{plot} is \code{FALSE}, the function prints summary statistics
and percentiles of the coverage data. If \code{plot} is \code{TRUE}, it prints a
histogram of the log-transformed coverage data.
}
\description{
This function calculates and optionally plots statistics for coverage data from
methylation sequencing experiments. It can handle both positional and regional
methylation data.
}
\examples{
\dontrun{
get_cov_stats(modseq_dat)
get_cov_stats(modseq_dat, plot = TRUE)
}

}
