% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mod_stats.R
\name{get_mod_stats}
\alias{get_mod_stats}
\title{Get Methylation Statistics from the ch3 Database}
\usage{
get_mod_stats(
  ch3_db,
  call_type = c("positions", "regions"),
  plot = FALSE,
  save_path = NULL
)
}
\arguments{
\item{ch3_db}{A string. The path to the database containing ch3 files from nanopore data.}

\item{call_type}{A character vector specifying the type of data to retrieve from the database. 
Default is "positions". Can also be "regions".}

\item{plot}{A logical value. If TRUE, a histogram of methylation values is plotted. Default is FALSE.}

\item{save_path}{Pathway to save the plot to. Usually .pdf or .png.}
}
\value{
If `plot` is FALSE, the function prints a summary of the methylation statistics and quantiles. 
If `plot` is TRUE, it displays a histogram of methylation values.
}
\description{
This function retrieves and calculates methylation statistics (mean methylation fractions) 
from a specified table in the ch3 database. It can either return summary statistics or plot 
a histogram of the methylation values, depending on the user's preference.
}
\details{
The function connects to the specified database, checks for the existence of the relevant table, 
and retrieves methylation fraction data. If the table contains methylation data, it prioritizes 
the `mh_frac` column over others. Depending on the `call_type`, it can compute statistics 
for either per base or per region. If `plot` is set to TRUE, it generates a histogram of the 
methylation values.
}
\note{
The function assumes that the database has tables named according to the `call_type` parameter 
(e.g., "positions", "regions"). It also expects specific columns for methylation data to exist.
}
\examples{
 # Specify the path to the database
 ch3_db <- system.file("my_data.ch3.db", package = "MethylseqR")
 
 # Get methylation statistics for the 'positions' call type without plotting
 get_mod_stats(ch3_db = ch3_db, call_type = "positions")
 
 get_mod_stats(ch3_db = ch3_db, call_type = "positions", plot = TRUE)

}
