% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_db.R
\name{open_dat}
\alias{open_dat}
\title{Open Methylation Data from a Directory}
\usage{
open_dat(
  path,
  samples = "all",
  chrs = "all",
  min_read_length = 100,
  min_call_prob = 0.9,
  min_base_qual = 10,
  max_memory = "8GB",
  max_threads = 4
)
}
\arguments{
\item{path}{Path to the database directory containing methylation data.}

\item{samples}{A vector of sample names to extract methylation data from.
Default is "all", which retrieves data from all available samples.}

\item{chrs}{A vector of chromosome names to extract. Default is "all",
which retrieves data from all available chromosomes.}

\item{min_read_length}{Minimum read length to allow. Filters out any reads
shorter than this length. Default is 100.}

\item{min_call_prob}{Minimum probability for base call. Default is 0.9.}

\item{min_base_qual}{Minimum Phred quality score for base calls. Default is 10.}

\item{max_memory}{Maximum memory allocation for the database connection.
Default is "8GB".}

\item{max_threads}{Maximum number of CPU threads to use for processing. Default is 4.}
}
\value{
A DuckDB object containing all methylation data that meets the specified
        filtering criteria.
}
\description{
This function opens methylation data stored in a specified directory and applies filters
based on sample names, chromosomes, read length, base quality, and call probability.
}
\examples{
data = open_dat("path/to/database")
data = open_dat("path/to/database", samples = c("sample1", "sample2"), chrs = c("chr1", "chr2"))

}
