% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_wrapper.R
\name{qc_wrapper}
\alias{qc_wrapper}
\title{Quality Control Wrapper for Methylation Data}
\usage{
qc_wrapper(ch3_db, call_type = "positions", plot = TRUE)
}
\arguments{
\item{ch3_db}{A database connection or object containing methylation data.}

\item{call_type}{A character string indicating the type of call to retrieve data (e.g., "positions", "regions").}

\item{plot}{A logical value indicating whether to generate plots for the statistical analyses.
Defaults to TRUE. Set to FALSE to suppress plots.}
}
\value{
This function does not return a value. It performs calculations and may produce plots based on the
        results of the analysis.
}
\description{
This function serves as a wrapper for various quality control analyses on methylation data.
It sequentially calculates coverage statistics, modification statistics, correlation analysis,
and performs principal component analysis (PCA).
}
\details{
The function first calculates coverage statistics using `get_cov_stats()`, then computes modification
         statistics with `get_mod_stats()`, follows up with correlation analysis via `cor_modseq()`, and finally
         runs principal component analysis with `pca_modseq()`.
}
\examples{
 # Specify the path to the database
 ch3_db <- system.file("my_data.ch3.db", package = "MethylSeqR")
 
 # Run quality control wrapper
 qc_wrapper(ch3_db, call_type = "positions")
  
}
