% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_positions.R
\name{summarize_positions}
\alias{summarize_positions}
\title{Summarize Methylation Positions from Calls Table}
\usage{
summarize_positions(
  ch3_db,
  mod_type = c("c", "m", "h", "mh"),
  chrs = c(paste0("chr", 1:22), paste0("Chr", 1:22), "chrX", "chrY", "chrM", "ChrX",
    "ChrY", "ChrM"),
  min_cov = 1
)
}
\arguments{
\item{ch3_db}{A DuckDB database connection or file path (character) to the `.ch3.db` file.}

\item{mod_type}{A character vector specifying the modification types to include. Options are  `"c"` (unmodified cytosine),
`"m"` (methylation), `"h"` (hydroxymethylation), 
  and `"mh"` (methylated + hydroxymethylated).}

\item{chrs}{A character vector specifying which chromosomes to include. Default includes all autosomes, 
sex chromosomes (chrX, chrY), and mitochondrial chromosome (chrM).}

\item{min_cov}{Minimum coverage required to include a position in the summary. Default is 1.}
}
\value{
The modified `ch3_db` object with the updated `positions` table.
}
\description{
This function processes methylation call data from a DuckDB database, summarizing 
coverage and call counts into a `positions` table.
}
\details{
The function:
- Connects to the DuckDB database.
- Summarizes methylation calls by `sample_name`, `chrom`, `start`, and `end`.
- Computes coverage and call counts (`m`, `h`, `mh`, `c`).
- Filters based on `min_cov`.
- Creates a `positions` table in the database.

A progress bar is displayed during execution.
}
\examples{
summarize_positions("example_ch3.ch3.db", min_cov = 1)

}
